% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_estimate_gxe.R
\name{ukb_estimate_gxe}
\alias{ukb_estimate_gxe}
\title{Estimate contribution of GRSxE to variance in outcome phenotype in UK Biobank}
\arguments{
\item{phenotype_name}{A string containing the column name of the phenotype to
use as outcome, e.g. '21001-0.0'}

\item{ukb_filename}{A string or character vector containing the path to the
file(s) containing the UK Biobank data}

\item{bgens_path}{path to the folder containing the \code{*.bgen} files with
the genetic data}

\item{snps}{a character vector of SNP RSIDs or a data.frame-type structure
containing a column named 'rsid' with SNP RSIDs, as well as an optional
column named 'beta' containing the effect estimates of the alternate allele
on the outcome}

\item{covariate_names}{a string or character vector containing the names of
any covariates to adjust for before estimating GxE}

\item{covariate_factor_names}{a string or character vector containing the
names of covariates to be considered as factors when adjusting for them}

\item{correct_age2_sex}{logical indicating whether to adjust for age, age^2,
and sex (in addition to any other covariates)}

\item{sample_ids}{a data.frame containing the individual ids under the column
name 'eid'. Any other included columns (e.g. genetic PCs) will be adjusted
for. Required if \code{sqc_filename} or \code{fam_filename} are missing}

\item{sqc_filename}{path to the UK Biobank sqc file (e.g. ukb_sqc_v2.txt).
Required if \code{sample_ids} is missing}

\item{fam_filename}{path to any \code{*.fam} file containing the sample ids
for the values in \code{sqc_filename}. Required if \code{sample_ids} is
missing}

\item{imp_sample_filename}{path to a \code{*.sample} file containing the
sample ids for the genetic data. The default is the first \code{*.sample}
in the \code{bgens_path} folder}

\item{ids_to_remove}{vector containing the ids of any samples to remove
before analysis}

\item{npcs}{number of genetic principal components to correct for before
analysis}

\item{betas}{vector of effects of SNPs on outcome. Ignored if \code{snps}
contains a column named 'beta'. Must be in the same order as \code{snps}}

\item{sim_num}{Number of permutations for bootstrap and fake GRSs}
}
\value{
\code{ukb_estimate_gxe} returns a list containing parameter estimates
  for alpha1, alpha2, beta, and gamma (\code{xopt}), their standard error
  (\code{SExopt}), and the associated p-values (\code{Pxopt}). The
  corresponding estimates for the fake GRS are stored in the \code{xopt0},
  \code{SExopt0}, and \code{Pxopt0}, respectively.

  In addition, the \code{Xopt} and \code{Xopt0} matrices contain the
  estimates for each bootstrap and fake GRS.

  The \code{tdiff} contains the t-statistic for the difference between the
  real data estimates and those from the fake GRS.
}
\description{
Estimate contribution of GRSxE to variance in outcome phenotype in UK Biobank
}
\examples{
\dontrun{
  library( GxE )
  # Load GWAS results from the Neale lab
  snps = get_betas_from_neale( neale_filename = '21001_irnt.gwas.imputed_v3.both_sexes.tsv.gz',
                               variants_filename = 'variants.tsv.gz' )

  # UK Biobank files not provided
  gxe  =  ukb_gxe_interaction( phenotype_name = '21001-0.0',
                               ukb_filename   = 'uk_biobank/pheno/ukb21067.csv',
                               bgens_path     = 'uk_biobank/imp',
                               snps           = snps,
                               sqc_filename   = 'uk_biobank/geno/ukb_sqc_v2.txt',
                               fam_filename   = 'uk_biobank/plink/ukb1638_cal_chr1_v2_s488366.fam' )

  print( gxe )
}
}
